#3 file upload - 28.06.2024  a panel of stories and an unfinished block
a panel with stories and an unfinished block for writing quick posts

The committee was created because of the instability of the electricity and laptop. A bug was noticed with the incorrect transfer of images from the main src/images folder to the dist/images folder - the bug is that the files are either not transferred or transferred but in the wrong format, and this is not on all photos, only on randomly selected ones. Most likely the problem is in the gulp configuration